+++ skyfield/__init__.py
@@ -5,5 +5,5 @@ functions and classes there.  See the ``documentation`` directory inside
 the repository, as well as the http://rhodesmill.org/skyfield/ site!
 
 """
-VERSION = 1, 49
+VERSION = 1, 50
 __version__ = '.'.join(map(str, VERSION))